refactor: Phase B - extract 12 medium-risk section components
Extract section-level components from 5 large parent files:

SettingsPage (1634→1341 lines):
- SettingsAppearanceSection: theme, timezone, view/sort settings
- SavedConversationsSection: conversation list with open/delete
- SettingsToolsSection: ImageProcessor embed + blocked users

AdminPage (1712→1386 lines):
- AdminUsersSection: user management table
- AdminSystemSection: API key, log level, version info
- AdminLogsSection: log viewer with filter/auto-refresh/export

CoinSearchChat (954→579 lines):
- CoinShowResultsGrid: show cards with calendar save
- CoinSuggestionGrid: suggestion cards with wishlist add

CoinDetailPage (1452→1082 lines):
- CoinDetailHeaderActions: back/purchase/sell/edit/delete
- CoinNumistaPanel: self-contained Numista search
- CoinActivityJournal: journal entries with add/delete

ImageProcessor (961→781 lines):
- ImageInputPanel: file drop zone + URL input

All components use typed props/emits. vue-tsc passes clean.
