+++ src/web/src/components/CoinSearchChat.vue
@@ -32,67 +32,23 @@
           </div>
 
           <!-- Coin Show results -->
-          <div v-if="msg.role === 'assistant' && msg.suggestions?.length && isCoinShowResults(msg.suggestions)" class="suggestions-grid">
-            <div v-for="(show, j) in (msg.suggestions as CoinShow[])" :key="j" class="show-card">
-              <div class="show-body">
-                <a v-if="show.url" :href="show.url" target="_blank" rel="noopener" class="show-name-link">
-                  <h4>{{ show.name }} <ExternalLink :size="12" /></h4>
-                </a>
-                <h4 v-else>{{ show.name }}</h4>
-                <div class="show-details">
-                  <span v-if="show.dates" class="show-detail"><Calendar :size="13" /> {{ show.dates }}</span>
-                  <span v-if="show.venue" class="show-detail"><MapPin :size="13" /> {{ show.venue }}</span>
-                  <span v-if="show.location" class="show-detail-sub">{{ show.location }}</span>
-                  <span v-if="show.entryFee" class="show-detail"><Ticket :size="13" /> {{ show.entryFee }}</span>
-                </div>
-                <p v-if="show.description" class="show-desc">{{ show.description }}</p>
-                <div v-if="show.notableDealers?.length" class="show-dealers">
-                  <span v-for="(dealer, k) in show.notableDealers" :key="k" class="meta-tag">{{ dealer }}</span>
-                </div>
-                <button
-                  class="save-cal-btn"
-                  :disabled="savedShows.has(showKey(show)) || savingShow === showKey(show)"
-                  @click="saveShowToCalendar(show)"
-                >
-                  <CalendarPlus :size="13" />
-                  {{ savedShows.has(showKey(show)) ? 'Saved' : savingShow === showKey(show) ? 'Saving...' : 'Save to Calendar' }}
-                </button>
-              </div>
-            </div>
-          </div>
+          <CoinShowResultsGrid
+            v-if="msg.role === 'assistant' && msg.suggestions?.length && isCoinShowResults(msg.suggestions)"
+            :shows="(msg.suggestions as CoinShow[])"
+            :saved-shows="savedShows"
+            :saving-show="savingShow"
+            @save-show="saveShowToCalendar"
+          />
 
           <!-- Coin suggestions after assistant message -->
-          <div v-if="msg.role === 'assistant' && msg.suggestions?.length && !isCoinShowResults(msg.suggestions)" class="suggestions-grid">
-            <div v-for="(coin, j) in (msg.suggestions as CoinSuggestion[])" :key="j" class="suggestion-card">
-              <div class="suggestion-img" v-if="getSuggestionImageUrl(coin)">
-                <img :src="getSuggestionImageUrl(coin)" :alt="coin.name" @error="handleImgError" />
-              </div>
-              <div class="suggestion-body">
-                <h4>{{ coin.name }}</h4>
-                <p class="suggestion-desc">{{ coin.description }}</p>
-                <div class="suggestion-meta">
-                  <span v-if="coin.era" class="meta-tag">{{ coin.era }}</span>
-                  <span v-if="coin.material" class="meta-tag">{{ coin.material }}</span>
-                  <span v-if="coin.denomination" class="meta-tag">{{ coin.denomination }}</span>
-                </div>
-                <div class="suggestion-price" v-if="coin.estPrice">{{ coin.estPrice }}</div>
-                <div class="suggestion-actions">
-                  <a v-if="coin.sourceUrl" :href="coin.sourceUrl" target="_blank" rel="noopener" class="source-link">
-                    <ExternalLink :size="12" /> {{ coin.sourceName || 'Source' }}
-                  </a>
-                  <button
-                    v-if="coin.era || coin.material || coin.denomination"
-                    class="btn btn-primary btn-sm add-btn"
-                    :disabled="addingIdx === `${i}-${j}`"
-                    @click="addToWishlist(coin, `${i}-${j}`)"
-                  >
-                    <CirclePlus :size="14" />
-                    {{ addedSet.has(`${i}-${j}`) ? 'Added!' : addingIdx === `${i}-${j}` ? 'Adding...' : 'Add to Wishlist' }}
-                  </button>
-                </div>
-              </div>
-            </div>
-          </div>
+          <CoinSuggestionGrid
+            v-if="msg.role === 'assistant' && msg.suggestions?.length && !isCoinShowResults(msg.suggestions)"
+            :suggestions="(msg.suggestions as CoinSuggestion[])"
+            :added-set="addedSet"
+            :adding-idx="addingIdx"
+            :message-index="i"
+            @add-to-wishlist="addToWishlist"
+          />
         </template>
 
         <div v-if="loading && !messages[messages.length-1]?.streaming" class="chat-bubble assistant">
@@ -117,13 +73,15 @@
 import { ref, nextTick, onMounted, onBeforeUnmount } from 'vue'
 import { agentChatStream, createCoin, proxyImage, scrapeImage, uploadImage, saveConversation, getPortfolioSummary, getAgentStatus, createCalendarEvent } from '@/api/client'
 import type { CoinSuggestion, CoinShow, AgentChatMessage, Category, Material } from '@/types'
-import { CirclePlus, ExternalLink, AlertTriangle, Calendar, MapPin, Ticket, CalendarPlus } from 'lucide-vue-next'
+import { AlertTriangle } from 'lucide-vue-next'
 import DOMPurify from 'dompurify'
 import { useDialog } from '@/composables/useDialog'
 import MarkdownIt from 'markdown-it'
 import ChatHeader from '@/components/chat/ChatHeader.vue'
 import ChatIntroPanel from '@/components/chat/ChatIntroPanel.vue'
 import ChatInputBar from '@/components/chat/ChatInputBar.vue'
+import CoinShowResultsGrid from '@/components/chat/CoinShowResultsGrid.vue'
+import CoinSuggestionGrid from '@/components/chat/CoinSuggestionGrid.vue'
 
 type ChatSuggestion = CoinSuggestion | CoinShow
 
@@ -467,49 +425,6 @@ async function saveShowToCalendar(show: CoinShow) {
   }
 }
 
-function proxyImageUrl(url: string): string {
-  if (!url) return ''
-  return `/api/proxy-image?url=${encodeURIComponent(url)}`
-}
-
-function getSuggestionImageUrl(coin: CoinSuggestion): string {
-  // Always prefer scraped image from sourceUrl (og:image is most reliable)
-  if (coin.sourceUrl) {
-    const cached = scrapedImages.value.get(coin.sourceUrl)
-    if (cached) return proxyImageUrl(cached)
-    if (cached === undefined) {
-      scrapedImages.value.set(coin.sourceUrl, '')
-      scrapeImage(coin.sourceUrl).then((res) => {
-        if (res.data.imageUrl) {
-          console.log('[agent] Scraped image from', coin.sourceUrl, '→', res.data.imageUrl)
-          scrapedImages.value.set(coin.sourceUrl, res.data.imageUrl)
-        } else if (coin.imageUrl) {
-          // Scrape returned nothing — fall back to agent-provided URL
-          console.log('[agent] Scrape empty, using agent imageUrl:', coin.imageUrl)
-          scrapedImages.value.set(coin.sourceUrl, coin.imageUrl)
-        }
-      }).catch(() => {
-        // Scrape failed — fall back to agent-provided URL
-        if (coin.imageUrl) {
-          console.log('[agent] Scrape failed, using agent imageUrl:', coin.imageUrl)
-          scrapedImages.value.set(coin.sourceUrl, coin.imageUrl)
-        }
-      })
-    }
-    return ''
-  }
-
-  // No sourceUrl — try agent imageUrl directly
-  if (coin.imageUrl) return proxyImageUrl(coin.imageUrl)
-  return ''
-}
-
-function handleImgError(e: Event) {
-  const img = e.target as HTMLImageElement
-  console.warn('[agent] Image failed to load:', img.src)
-  img.style.display = 'none'
-}
-
 onMounted(async () => {
   inputBarEl.value?.focus()
   if (props.loadConversation) {
@@ -722,233 +637,9 @@ function handleViewportResize() {
   50% { opacity: 0; }
 }
 
-.suggestions-grid {
-  display: flex;
-  flex-direction: column;
-  gap: 0.6rem;
-  width: 100%;
-}
-
-.suggestion-card {
-  background: var(--bg-card);
-  border: 1px solid var(--border-subtle);
-  border-radius: var(--radius-md);
-  overflow: hidden;
-  display: flex;
-  transition: border-color var(--transition-fast);
-}
-
-.suggestion-card:hover {
-  border-color: var(--accent-gold);
-}
-
-.suggestion-img {
-  width: 80px;
-  min-height: 80px;
-  flex-shrink: 0;
-  overflow: hidden;
-  background: var(--bg-body);
-}
-
-.suggestion-img img {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-}
-
-.suggestion-body {
-  padding: 0.6rem 0.75rem;
-  flex: 1;
-  min-width: 0;
-}
-
-.suggestion-body h4 {
-  font-size: 0.85rem;
-  margin: 0 0 0.25rem;
-  color: var(--text-primary);
-  line-height: 1.3;
-}
-
-.suggestion-desc {
-  font-size: 0.78rem;
-  color: var(--text-secondary);
-  margin: 0 0 0.4rem;
-  display: -webkit-box;
-  -webkit-line-clamp: 2;
-  -webkit-box-orient: vertical;
-  overflow: hidden;
-}
-
-.suggestion-meta {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 0.3rem;
-  margin-bottom: 0.4rem;
-}
-
-.meta-tag {
-  font-size: 0.7rem;
-  padding: 0.1rem 0.4rem;
-  border-radius: var(--radius-full);
-  background: var(--bg-body);
-  color: var(--text-muted);
-  border: 1px solid var(--border-subtle);
-}
-
-.suggestion-price {
-  font-size: 0.82rem;
-  font-weight: 600;
-  color: var(--accent-gold);
-  margin-bottom: 0.4rem;
-}
-
-.suggestion-actions {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 0.5rem;
-}
-
-.source-link {
-  font-size: 0.72rem;
-  color: var(--text-muted);
-  text-decoration: none;
-  display: flex;
-  align-items: center;
-  gap: 0.2rem;
-  transition: color var(--transition-fast);
-}
-
-.source-link:hover {
-  color: var(--accent-gold);
-}
-
-/* Coin Show cards */
-.show-card {
-  background: var(--bg-card);
-  border: 1px solid var(--border-subtle);
-  border-radius: var(--radius-md);
-  overflow: hidden;
-  transition: border-color var(--transition-fast);
-}
-
-.show-card:hover {
-  border-color: var(--accent-gold);
-}
-
-.show-body {
-  padding: 0.7rem 0.85rem;
-}
-
-.show-body h4 {
-  font-size: 0.88rem;
-  margin: 0 0 0.4rem;
-  color: var(--text-primary);
-  line-height: 1.3;
-}
-
-.show-name-link {
-  text-decoration: none;
-  color: inherit;
-}
-
-.show-name-link h4 {
-  display: flex;
-  align-items: center;
-  gap: 0.35rem;
-  color: var(--accent-gold);
-  transition: color var(--transition-fast);
-}
-
-.show-name-link:hover h4 {
-  color: var(--accent-bronze);
-}
-
-.show-details {
-  display: flex;
-  flex-direction: column;
-  gap: 0.25rem;
-  margin-bottom: 0.4rem;
-}
-
-.show-detail {
-  display: flex;
-  align-items: center;
-  gap: 0.35rem;
-  font-size: 0.8rem;
-  color: var(--text-secondary);
-}
-
-.show-detail-sub {
-  font-size: 0.75rem;
-  color: var(--text-muted);
-  padding-left: 1.5rem;
-}
-
-.show-desc {
-  font-size: 0.78rem;
-  color: var(--text-secondary);
-  margin: 0 0 0.4rem;
-  display: -webkit-box;
-  -webkit-line-clamp: 2;
-  -webkit-box-orient: vertical;
-  overflow: hidden;
-}
-
-.show-dealers {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 0.25rem;
-}
-
-.save-cal-btn {
-  display: flex;
-  align-items: center;
-  gap: 0.35rem;
-  font-size: 0.75rem;
-  padding: 0.3rem 0.65rem;
-  margin-top: 0.5rem;
-  background: transparent;
-  border: 1px solid var(--border-subtle);
-  border-radius: var(--radius-sm);
-  color: var(--text-secondary);
-  cursor: pointer;
-  transition: all var(--transition-fast);
-  font-family: inherit;
-}
-
-.save-cal-btn:hover:not(:disabled) {
-  border-color: var(--accent-gold-dim);
-  color: var(--accent-gold);
-  background: var(--accent-gold-glow);
-}
-
-.save-cal-btn:disabled {
-  opacity: 0.6;
-  cursor: default;
-}
-
-.add-btn {
-  display: flex;
-  align-items: center;
-  gap: 0.3rem;
-  font-size: 0.72rem;
-  padding: 0.3rem 0.6rem;
-  flex-shrink: 0;
-}
-
 @media (max-width: 640px) {
   .chat-drawer {
     width: 100%;
-  }
-
-  .suggestion-card {
-    flex-direction: column;
-  }
-
-  .suggestion-img {
-    width: 100%;
-    height: 120px;
   }
 }
 </style>
