+++ src/web/src/components/ImageProcessor.vue
@@ -1,25 +1,15 @@
 <template>
   <div class="image-processor">
     <!-- Step 1: Image Input -->
-    <div v-if="!sourceImage" class="input-section card">
-      <h3>Load Image</h3>
-      <div class="input-methods">
-        <label class="drop-zone" :class="{ dragging }" @dragover.prevent="dragging = true"
-          @dragleave="dragging = false" @drop.prevent="handleDrop">
-          <Upload :size="32" />
-          <span>Drop an image here or click to browse</span>
-          <input type="file" accept="image/*" hidden @change="handleFileSelect" />
-        </label>
-        <div class="url-input-row">
-          <input v-model="imageUrl" type="url" class="form-input" placeholder="Or paste an image URL..."
-            @keydown.enter="handleUrlLoad" />
-          <button class="btn btn-primary btn-sm" :disabled="!imageUrl || urlLoading" @click="handleUrlLoad">
-            {{ urlLoading ? 'Loading...' : 'Fetch' }}
-          </button>
-        </div>
-        <p v-if="inputError" class="msg error">{{ inputError }}</p>
-      </div>
-    </div>
+    <ImageInputPanel
+      v-if="!sourceImage"
+      :source-image="sourceImage"
+      :url-loading="urlLoading"
+      :input-error="inputError"
+      @file-select="loadImageFromFile"
+      @url-load="handleUrlLoad"
+      @drop="loadImageFromFile"
+    />
 
     <!-- Step 2: Processing & Preview -->
     <div v-if="sourceImage" class="processing-section">
@@ -137,10 +127,10 @@
 <script setup lang="ts">
 import { ref, watch, nextTick, onMounted, onUnmounted } from 'vue'
 import { removeBackground as removeBg } from '@imgly/background-removal'
-import { Upload } from 'lucide-vue-next'
 import { proxyImage, uploadImage, createCoin, getCoins } from '@/api/client'
 import { getCoin } from '@/api/client'
 import type { Coin } from '@/types'
+import ImageInputPanel from '@/components/ImageInputPanel.vue'
 
 const props = defineProps<{
   coinId?: number
@@ -152,10 +142,8 @@ const emit = defineEmits<{
 
 // Input state
 const sourceImage = ref<string | null>(null)
-const imageUrl = ref('')
 const urlLoading = ref(false)
 const inputError = ref('')
-const dragging = ref(false)
 
 // Processing state
 type Step = 'preview' | 'removing' | 'crop' | 'done'
@@ -217,23 +205,12 @@ function loadImageFromFile(file: File) {
   reader.readAsDataURL(file)
 }
 
-function handleFileSelect(e: Event) {
-  const file = (e.target as HTMLInputElement).files?.[0]
-  if (file) loadImageFromFile(file)
-}
-
-function handleDrop(e: DragEvent) {
-  dragging.value = false
-  const file = e.dataTransfer?.files?.[0]
-  if (file) loadImageFromFile(file)
-}
-
-async function handleUrlLoad() {
-  if (!imageUrl.value) return
+async function handleUrlLoad(url: string) {
+  if (!url) return
   inputError.value = ''
   urlLoading.value = true
   try {
-    const res = await proxyImage(imageUrl.value)
+    const res = await proxyImage(url)
     const blob = res.data as Blob
     const reader = new FileReader()
     reader.onload = (e) => {
@@ -611,47 +588,6 @@ onUnmounted(() => {
 .image-processor {
   max-width: 700px;
   margin: 0 auto;
-}
-
-.input-section h3 {
-  margin-bottom: 1rem;
-}
-
-.input-methods {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.drop-zone {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  gap: 0.75rem;
-  padding: 2.5rem 1.5rem;
-  border: 2px dashed var(--border-subtle);
-  border-radius: var(--radius-md);
-  color: var(--text-muted);
-  cursor: pointer;
-  transition: all var(--transition-fast);
-  text-align: center;
-}
-
-.drop-zone:hover,
-.drop-zone.dragging {
-  border-color: var(--accent-gold);
-  color: var(--accent-gold);
-  background: var(--accent-gold-glow);
-}
-
-.url-input-row {
-  display: flex;
-  gap: 0.5rem;
-}
-
-.url-input-row .form-input {
-  flex: 1;
 }
 
 /* Steps */
