+++ src/web/src/pages/AdminPage.vue
@@ -22,45 +22,14 @@
       </div>
 
       <!-- Users Tab -->
-      <section v-if="activeTab === 'users'" class="admin-section card">
-        <h2>User Management</h2>
-        <div v-if="usersLoading" class="loading-overlay"><div class="spinner"></div></div>
-        <table v-else class="users-table">
-          <thead>
-            <tr>
-              <th>Username</th>
-              <th>Role</th>
-              <th>Created</th>
-              <th>Actions</th>
-            </tr>
-          </thead>
-          <tbody>
-            <tr v-for="user in users" :key="user.id">
-              <td>
-                <span class="username">{{ user.username }}</span>
-                <span v-if="user.id === auth.user?.id" class="you-badge">(you)</span>
-              </td>
-              <td>
-                <span class="badge" :class="`badge-${user.role === 'admin' ? 'roman' : 'modern'}`">
-                  {{ user.role }}
-                </span>
-              </td>
-              <td class="date-cell">{{ formatDate(user.createdAt) }}</td>
-              <td>
-                <div v-if="user.id !== auth.user?.id" class="action-btns">
-                  <button class="btn btn-secondary btn-sm" @click="openResetModal(user)">
-                    Reset
-                  </button>
-                  <button class="btn btn-danger btn-sm" @click="handleDeleteUser(user)">
-                    Delete
-                  </button>
-                </div>
-                <span v-else class="text-muted">—</span>
-              </td>
-            </tr>
-          </tbody>
-        </table>
-      </section>
+      <AdminUsersSection
+        v-if="activeTab === 'users'"
+        :users="users"
+        :loading="usersLoading"
+        :current-user-id="auth.user?.id ?? 0"
+        @reset="openResetModal"
+        @delete="handleDeleteUser"
+      />
 
       <!-- AI Tab -->
       <section v-if="activeTab === 'ai'" class="admin-section card">
@@ -271,75 +240,31 @@
       </section>
 
       <!-- System Tab -->
-      <section v-if="activeTab === 'system'" class="admin-section card">
-        <h2>System Settings</h2>
-        <form @submit.prevent="saveSettings">
-          <div class="form-group">
-            <label class="form-label">Numista API Key</label>
-            <input v-model="settings.NumistaAPIKey" class="form-input" type="password" placeholder="Enter your Numista API key" />
-            <span class="form-hint">Get a free key at <a href="https://en.numista.com/api/" target="_blank" rel="noopener">numista.com/api</a> (2,000 requests/month free)</span>
-          </div>
-          <div class="form-group">
-            <label class="form-label">Log Level</label>
-            <select v-model="settings.LogLevel" class="form-select">
-              <option v-for="level in LOG_LEVELS" :key="level" :value="level">{{ level }}</option>
-            </select>
-          </div>
-          <p v-if="settingsMsg" class="msg" :class="{ error: settingsError }">{{ settingsMsg }}</p>
-          <button type="submit" class="btn btn-primary btn-sm" :disabled="settingsSaving">
-            {{ settingsSaving ? 'Saving...' : 'Save System Settings' }}
-          </button>
-        </form>
-        <div class="version-info">
-          <span class="version-label">Version</span>
-          <span class="version-value">{{ appVersion }}</span>
-          <span v-if="buildDate" class="version-date">Built {{ buildDate }}</span>
-        </div>
-      </section>
+      <AdminSystemSection
+        v-if="activeTab === 'system'"
+        :numista-api-key="settings.NumistaAPIKey ?? ''"
+        :log-level="settings.LogLevel ?? ''"
+        :log-levels="LOG_LEVELS"
+        :saving="settingsSaving"
+        :msg="settingsMsg"
+        :error="settingsError"
+        :app-version="appVersion"
+        :build-date="buildDate"
+        @save="onSystemSave"
+      />
 
       <!-- Logs Tab -->
-      <section v-if="activeTab === 'logs'" class="admin-section card">
-        <h2>Application Logs</h2>
-        <div class="logs-toolbar">
-          <select v-model="logsFilter" class="form-select logs-filter" @change="loadLogs">
-            <option value="">All Levels</option>
-            <option v-for="level in ['TRACE','DEBUG','INFO','WARN','ERROR']" :key="level" :value="level">{{ level }}</option>
-          </select>
-          <button class="btn btn-secondary btn-sm" @click="loadLogs" :disabled="logsLoading">
-            {{ logsLoading ? 'Loading...' : 'Refresh' }}
-          </button>
-          <button
-            class="btn btn-sm"
-            :class="logsAutoRefresh ? 'btn-primary' : 'btn-secondary'"
-            @click="toggleAutoRefresh"
-          >
-            {{ logsAutoRefresh ? 'Auto ●' : 'Auto ○' }}
-          </button>
-          <button
-            class="btn btn-secondary btn-sm"
-            @click="exportLogs"
-            :disabled="logs.length === 0"
-            title="Export logs as text file"
-          >
-            <Download :size="14" /> Export
-          </button>
-        </div>
-        <div class="logs-container">
-          <div v-if="logs.length === 0 && !logsLoading" class="logs-empty">
-            No log entries. Click Refresh to load.
-          </div>
-          <div
-            v-for="(entry, i) in logs"
-            :key="i"
-            class="log-entry"
-            :class="logLevelClass(entry.level)"
-          >
-            <span class="log-time">{{ entry.timestamp.substring(11, 19) }}</span>
-            <span class="log-level-badge">{{ entry.level }}</span>
-            <span class="log-msg">{{ entry.message }}</span>
-          </div>
-        </div>
-      </section>
+      <AdminLogsSection
+        v-if="activeTab === 'logs'"
+        :logs="logs"
+        :loading="logsLoading"
+        :filter="logsFilter"
+        :auto-refresh="logsAutoRefresh"
+        @load="loadLogs"
+        @toggle-auto-refresh="toggleAutoRefresh"
+        @export="exportLogs"
+        @update:filter="logsFilter = $event"
+      />
 
       <!-- Schedules Tab -->
       <section v-if="activeTab === 'schedules'" class="admin-section card">
@@ -621,6 +546,9 @@ import { LOG_LEVELS } from '@/types'
 import type { UserInfo, AppSettings, LogEntry, AvailabilityRun, ValuationRun } from '@/types'
 import { useDialog } from '@/composables/useDialog'
 import ResetPasswordModal from '@/components/admin/ResetPasswordModal.vue'
+import AdminUsersSection from '@/components/admin/AdminUsersSection.vue'
+import AdminSystemSection from '@/components/admin/AdminSystemSection.vue'
+import AdminLogsSection from '@/components/admin/AdminLogsSection.vue'
 import { Users, Cpu, Wrench, ScrollText, Download, ShieldCheck, CalendarClock } from 'lucide-vue-next'
 
 const tabIcons: Record<string, Component> = { users: Users, ai: Cpu, system: Wrench, logs: ScrollText, schedules: CalendarClock }
@@ -885,18 +813,14 @@ function exportLogs() {
   URL.revokeObjectURL(url)
 }
 
-function logLevelClass(level: string) {
-  switch (level) {
-    case 'ERROR': return 'log-error'
-    case 'WARN': return 'log-warn'
-    case 'DEBUG': return 'log-debug'
-    case 'TRACE': return 'log-trace'
-    default: return 'log-info'
-  }
-}
-
 function formatDate(dateStr: string) {
   return new Date(dateStr).toLocaleDateString()
+}
+
+function onSystemSave(payload: { numistaApiKey: string; logLevel: string }) {
+  settings.value.NumistaAPIKey = payload.numistaApiKey
+  settings.value.LogLevel = payload.logLevel
+  saveSettings()
 }
 
 // Availability
@@ -1119,24 +1043,9 @@ onUnmounted(() => {
   font-weight: 600;
 }
 
-.username {
-  font-weight: 500;
-}
-
-.you-badge {
-  font-size: 0.7rem;
-  color: var(--text-muted);
-  margin-left: 0.3rem;
-}
-
 .date-cell {
   font-size: 0.85rem;
   color: var(--text-secondary);
-}
-
-.action-btns {
-  display: flex;
-  gap: 0.4rem;
 }
 
 .text-muted {
@@ -1307,103 +1216,14 @@ onUnmounted(() => {
   .users-table {
     font-size: 0.85rem;
   }
-  .action-btns {
-    flex-direction: column;
-  }
 }
 
 /* Logs */
-.logs-toolbar {
-  display: flex;
-  gap: 0.5rem;
-  align-items: center;
-  margin-bottom: 1rem;
-}
-
-.logs-filter {
-  width: auto;
-  min-width: 120px;
-}
-
-.logs-container {
-  max-height: 500px;
-  overflow-y: auto;
-  background: var(--bg-body);
-  border: 1px solid var(--border-subtle);
-  border-radius: var(--radius-sm);
-  padding: 0.5rem;
-  font-family: 'Courier New', Courier, monospace;
-  font-size: 0.78rem;
-  line-height: 1.5;
-}
-
 .logs-empty {
   text-align: center;
   padding: 2rem;
   color: var(--text-muted);
   font-family: 'Inter', sans-serif;
-}
-
-.log-entry {
-  display: flex;
-  gap: 0.5rem;
-  padding: 0.15rem 0.25rem;
-  border-radius: 2px;
-}
-
-.log-entry:hover {
-  background: var(--bg-card);
-}
-
-.log-time {
-  color: var(--text-muted);
-  flex-shrink: 0;
-}
-
-.log-level-badge {
-  flex-shrink: 0;
-  min-width: 48px;
-  text-align: center;
-  font-weight: 600;
-  border-radius: 2px;
-  padding: 0 4px;
-}
-
-.log-msg {
-  word-break: break-word;
-}
-
-.log-error .log-level-badge { color: #e74c3c; }
-.log-error .log-msg { color: #e74c3c; }
-.log-warn .log-level-badge { color: #f39c12; }
-.log-debug .log-level-badge { color: #3498db; }
-.log-trace .log-level-badge { color: #7f8c8d; }
-.log-info .log-level-badge { color: #2ecc71; }
-
-.version-info {
-  margin-top: 1.5rem;
-  padding-top: 1rem;
-  border-top: 1px solid var(--border-subtle);
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  font-size: 0.78rem;
-  color: var(--text-muted);
-}
-
-.version-label {
-  font-weight: 600;
-  text-transform: uppercase;
-  letter-spacing: 0.05em;
-}
-
-.version-value {
-  font-family: 'Courier New', Courier, monospace;
-  color: var(--text-secondary);
-}
-
-.version-date {
-  margin-left: 0.25rem;
 }
 
 /* Availability */
